Add timings to TkInterface class documentation
Timed the instantiation of the Tk interface, which is quite slow (more than a second on a modern system), and moreover, adds on top of the VM startup time.

This was an investigation of the feasibility of converting the client/server app design to a simple monolith, which is not possible given the slowness.